fixed rate products, order info, categories

+++ src/main/java/org/shop/sportwebstore/model/ProductInOrder.java
@@ -4,16 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductInOrder {
     private String productId;
     private int amount;
-    private double price;
+    private BigDecimal price;
     private boolean isRated;
 
-    public ProductInOrder(String productId, int amount, double price) {
+    public ProductInOrder(String productId, int amount, BigDecimal price) {
         this.productId = productId;
         this.amount = amount;
         this.price = price;
